+++ dali/image/CMakeLists.txt
@@ -1,4 +1,4 @@
-# Copyright (c) 2017-2018, NVIDIA CORPORATION. All rights reserved.
+# Copyright (c) 2017-2022, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -22,3 +22,8 @@ if (NOT BUILD_LIBTIFF)
         ${DALI_SRC_DIR}/dali/image/tiff_libtiff.cc
     )
 endif()
+
+set(DALI_SRCS ${DALI_SRCS}
+    "${DALI_ROOT}/dali/imgcodec/decoders/jpeg/jpeg_mem.cc"
+    "${DALI_ROOT}/dali/imgcodec/decoders/jpeg/jpeg_handle.cc"
+    PARENT_SCOPE)
