+++ dali/imgcodec/decoders/jpeg/jpeg_handle.h
@@ -18,12 +18,12 @@ limitations under the License.
 // This file declares the functions and structures for memory I/O with libjpeg
 // These functions are not meant to be used directly, see jpeg_mem.h instead.
 
-#ifndef DALI_IMAGE_JPEG_HANDLE_H_
-#define DALI_IMAGE_JPEG_HANDLE_H_
+#ifndef DALI_IMGCODEC_DECODERS_JPEG_JPEG_HANDLE_H_
+#define DALI_IMGCODEC_DECODERS_JPEG_JPEG_HANDLE_H_
 
 #include <string>
 #include "dali/core/common.h"
-#include "dali/image/jpeg_utils.h"
+#include "dali/imgcodec/decoders/jpeg/jpeg_utils.h"
 
 namespace dali {
 namespace jpeg {
@@ -60,4 +60,4 @@ void SetDest(j_compress_ptr cinfo, void *buffer, int bufsize,
 }  // namespace jpeg
 }  // namespace dali
 
-#endif  // DALI_IMAGE_JPEG_HANDLE_H_
+#endif  // DALI_IMGCODEC_DECODERS_JPEG_JPEG_HANDLE_H_
