+++ dali/imgcodec/decoders/jpeg/jpeg_mem.h
@@ -20,13 +20,13 @@ limitations under the License.
 // (data array and size fields).
 // Direct manipulation of JPEG strings are supplied: Flip, Rotate, Crop..
 
-#ifndef DALI_IMAGE_JPEG_MEM_H_
-#define DALI_IMAGE_JPEG_MEM_H_
+#ifndef DALI_IMGCODEC_DECODERS_JPEG_JPEG_MEM_H_
+#define DALI_IMGCODEC_DECODERS_JPEG_JPEG_MEM_H_
 
 #include <functional>
 #include <string>
 #include "dali/core/common.h"
-#include "dali/image/jpeg_utils.h"
+#include "dali/imgcodec/decoders/jpeg/jpeg_utils.h"
 
 namespace dali {
 namespace jpeg {
@@ -162,4 +162,4 @@ bool Compress(const void* srcdata, int width, int height,
 }  // namespace jpeg
 }  // namespace dali
 
-#endif  // DALI_IMAGE_JPEG_MEM_H_
+#endif  // DALI_IMGCODEC_DECODERS_JPEG_JPEG_MEM_H_
